requirements.txt
aloe>=0.2.0,<1.0.0
aloe-django>=0.2.0,<1.0.0
ansicolors>=1.1.8,<1.2.0
asgiref>=3.2.7,<3.3.0
colorama>=0.4.3,<1.0.0
cssselect>=1.1.0,<1.2.0
Django>=3.0.3,<3.1.0
django-nose>=1.4.6,<1.5.0
djangorestframework>=3.11.0,<3.12.0
gherkin-official>=4.1.3,<4.2.0
nose>=1.3.7,<1.4.0
psycopg2>=2.8.5,<2.9.0
pyasn1>=0.4.8,<1.0.0
pyasn1-modules>=0.2.8,<1.0.0
pytz==2019.3
PyYAML>=5.3.1,<5.4.0
sqlparse>=0.3.1,<1.0.0
Pygments>=2.6.1,<2.7.0
flake8>=3.8.2,<3.9.0